feat(service-provider): 实现服务商管理功能并完善国家指南逻辑
- 新增CountryGuideServiceImpl服务实现类
- 修改ServiceProvider实体字段类型，将isAbroad从Integer改为Byte
- 更新数据库schema，调整逻辑删除字段默认值并优化索引
- 重构CommonController.getCountries方法，从CountryGuide表动态获取国家列表
- 修复前端企业页面的标签处理逻辑和UI显示问题
- 更新API文档，修正isAbroad字段类型说明

+++ zhilian-backend/src/main/java/com/zhilian/zhilianbackend/entity/ServiceProvider.java
@@ -135,7 +135,7 @@ public class ServiceProvider implements Serializable {
      * 是否提供出海服务（0否 1是）
      */
     @TableField("is_abroad")
-    private Integer isAbroad;
+    private Byte isAbroad;
 
     /**
      * 覆盖国家/地区，多个用逗号分隔（如"欧盟,美国"）
